edited customersql, delete, home, user for delete button

+++ home.php
@@ -84,7 +84,9 @@
                             echo "<div class='card-title'>";
                                 echo "<a href='user.php?username=$uname'>".$uname."</a>";
                                 if($uname==$name){
-                                    echo "<button type='button' class='btn btn-danger float-right' data-toggle='modal' data-target='#delete'><i class='fa fa-trash'></i></button>";
+                                    echo "<a class='btn btn-danger float-right' href='delete.php?id=".$row1['post_id']."&image=".$row1['gambar']."' 
+                            onclick='return confirm(\"Are you sure you want to delete this movie?\")'><i class='fa fa-trash'></i> Delete</a></td>";
+                                    //echo "<button type='button' class='btn btn-danger float-right' data-toggle='modal' data-target='#delete'><i class='fa fa-trash'></i></button>";
                                 }       
                                 echo "</div>";                                  
                             echo "<img src='".$picture."'/ width='600px;' class='img-fluid'>";
